phaser linked up with Tiled, for creating maps and attack/state/animation set up for Tank.
i have setup Tiled a program for using tilesets and creating the map or maps. i have added a property to every tile called collides set to boolean (false). this is done in Tiled but is also added in phaser so phaser will use this property and add physics. the Tank character is now able to attack and has an attack state.